Fixed Item's metadata's mapping's amount as double.

+++ app/src/main/java/org/lequochai/fashionshop/ItemDetailActivity.java
@@ -111,7 +111,7 @@ public class ItemDetailActivity extends AppCompatActivity {
         }
 
         lblPrice.setText("Giá: " + mapping.get("price"));
-        lblAmount.setText("Số lượng: " + mapping.get("amount"));
+        lblAmount.setText("Số lượng: " + (int)((double)mapping.get("amount")));
     }
 
 //    Methods:
